pyproject.toml
[project]
name = "warpspeed25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai[tools]>=0.130.0",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
